requirements.txt
matplotlib==3.6.0
numpy==1.23.3
pandas==1.5.0
Pillow==9.2.0
requests==2.28.1
streamlit==1.12.2
